[rules] reorganize rules, matchers, and helpers
