+++ rules/community/cloudtrail/cloudtrail_put_bucket_acl.py
