+++ rules/community/cloudtrail/cloudtrail_put_object_acl.py
