+++ rules/community/cloudtrail/cloudtrail_root_account.py
